chore: Moves useOnClickOutside to fs/component hooks and replace
