+++ packages/components/src/hooks/index.ts
@@ -1,6 +1,6 @@
 export { default as UIProvider, Toast as ToastProps, useUI } from './UIProvider'
 export { useFadeEffect } from './useFadeEffect'
-export { useTrapFocus } from './useTrapFocus'
+export { useOnClickOutside } from './useOnClickOutside'
 export { useSearch } from './useSearch'
 export { useSKUMatrix } from './useSKUMatrix'
 export { useScrollDirection } from './useScrollDirection'
@@ -12,3 +12,4 @@ export type {
   SlideDirection,
 } from './useSlider'
 export { useSlideVisibility } from './useSlideVisibility'
+export { useTrapFocus } from './useTrapFocus'
